Added system to ignore certain file paths.
The config file now can optionally have an "ignore" section containing
file name patterns that the linter will ignore. The patterns follow glob
like syntax (using fnmatch under the hood).

An example of a configuration is below.

ignore:
  - "*/migrations/*"

This is useful when files are generated by a framework that will not pass
the linter and you wish to always ignore them.